Finalizing. (Pls run migration on all apps)
There have been updates to the models, so please do as the following
-Run the following commands in terminal or bash
- Make migrations for all apps
# $ python manage.py makemigrations about_us accounts blogs core dashboard documents news suppliers task_manager vacancies visit_counter

- While you  run the above command, there will be prompts asking if you changed the model names  or asking you to define first time values.
For the model name updates, type 'yes' to confirm that you want to update the model names.
For the field name updates, type 'yes' to confirm that you want to update the field names.
For the default field values:
   - type 1 if the field is a foreign key value
   - type timezone.now() if the field is a date field

After that apply the migration
$ python manage.py migrate

Then run the following
$python manage.py update_translation_fields